refactor(core): Extract Module + BDD tests for code review compliance
NLOC extraction (Martin Fowler Extract Module):
- Extract match_dispatch.rs from select_dispatch.rs (513→<500 lines)
  Contains: dispatch_match_query, execute_match_pipeline/parallel,
  finalize_match_results, merge_match_results, MATCH_METRICS static
- Extract join_pushdown.rs from query_engine.rs (547→<500 lines)
  Contains: analyze_join_pushdown_for_select, apply_post_join_filters

BDD tests for VectorFirst MATCH strategy (bdd-testing.md compliance):
- test_match_vector_first_basic: nominal similarity + graph pattern
- test_match_vector_first_filters_by_graph: high-sim without edges excluded
- test_match_vector_first_empty_collection: edge case empty results
- test_match_vector_first_high_threshold: edge case threshold eliminates all
- test_match_vector_first_missing_param: negative missing $query param
- test_match_vector_first_no_similarity: negative falls back to GraphFirst
